requirements.txt
textual 
stem
pysocks
rsa